- Removed any 64-bit logic in kernel (to avoid hitting 64-bit bugs)

+++ Source/Mosa.Kernel.x86/ConsoleSession.cs
@@ -244,10 +244,10 @@ namespace Mosa.Kernel.x86
 		/// <param name="val">The val.</param>
 		/// <param name="digits">The digits.</param>
 		/// <param name="size">The size.</param>
-		public void Write(ulong val, byte digits, int size)
+		public void Write(uint val, byte digits, int size)
 		{
 			uint count = 0;
-			ulong temp = val;
+			uint temp = val;
 
 			do
 			{
